feat: ComponentCollectionViewEvent enum에 DidSelect notificationName과 post 액션 추가
- collectionView의 cell이 선택되었을 때를 사용할 Notification

#36
